Replace wgetASF query with ESA query for orbits
These updates replace the former wgetasf alias command with the language from the ESA API system to automatically download precise orbit files for the given SAFE file names specified by the user. The logic has been updated to query the ESA database for a specific date of file and download that file, instead of querying the entire database for a single list of all files and finding the matching file on the list. I also updated the $orb variable to be $orbit. In addition, I updated the usage output to reflect these changes.