Update input call in extract_one_time_series.csh
In line 64, it called "scene.tab" by exact name instead of the user-input scene.tab. This replaces the name call with the $5 variable call, so the user-specified scene.tab file name is called (i.e. this introduces more flexibility if a user's scene.tab file is called something slightly different).