Update correct_insar_with_gnss.csh for upsampling
Instead of using surface to up-sample the filtered correction grid, I replace it with grdsample. This should help avoid systematic fluctuations in the gridded final product. The surface command is left commented out in case a user wants to try it. Additionally, I remove the redundant grdsample between tmp2 and tmp3 grids.