Fix issue with two lines being passed to esaID
In cases where two possible orbit files fit the query, two esaIDs get passed to the ${esaID} variable, which breaks the if statement. I put an if statement to check if there are two lines in the esaID file. If there are, we want the later one (the last one), so I have it print only the second line, which matches the tail -1 command on the orbit file name. I added this in both the restituted and precise orbit file calls.